댓글 관련 int -> long 변경 메서드 수정

+++ src/main/java/com/yh/web/service/CommentService.java
@@ -8,11 +8,11 @@ public interface CommentService {
     //댓글 10개씩 보여주기
     int listNum = 10;
 
-    Map<String, Object> selectCommentByArticleNo(int articleNo, int page);
+    Map<String, Object> selectCommentByArticleNo(long articleNo, long page);
 
-    Map<String, Object> selectCommentByCno(int cno, int page);
+    Map<String, Object> selectCommentByCno(long cno, long page);
 
     int addComment(Comment comment);
 
-    int deleteComment(int cno);
+    int deleteComment(long cno);
 }
